Extract feed module (again)
- beef: doesn't know about FilmDetail so it can't start that activity
