+++ app/src/main/java/com/muvi/MuViApplication.kt
@@ -4,13 +4,12 @@ import android.app.Application
 import org.koin.android.ext.koin.androidContext
 import org.koin.core.context.startKoin
 
-class MuViApplication: Application() {
+class MuViApplication : Application() {
 
     override fun onCreate() {
         super.onCreate()
         startKoin {
             androidContext(this@MuViApplication)
-            modules(muviModule)
         }
     }
 }
