+++ feed/src/main/java/com/muvi/feed/FeedFragment.kt
@@ -1,5 +1,4 @@
-package com.muvi
-
+package com.muvi.feed
 
 import android.content.Intent
 import android.os.Bundle
@@ -8,13 +7,20 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.Observer
+import com.muvi.FeedViewModel
 import com.muvi.domain.Film
 import kotlinx.android.synthetic.main.fragment_feed.*
 import org.koin.androidx.viewmodel.ext.android.viewModel
+import org.koin.core.context.loadKoinModules
 
 internal class FeedFragment : Fragment() {
 
     private val feedViewModel: FeedViewModel by viewModel()
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        loadKoinModules(feedModule)
+    }
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
         return inflater.inflate(R.layout.fragment_feed, container, false)
