+++ settings.gradle
@@ -1,6 +1,7 @@
 include ':app'
 include ':data'
 include ':domain'
+include ':feed'
 include ':film_detail'
 include ':film_detail_data'
 include ':film_detail_domain'
